Initial working implemention of DeviceAdapter using cordova-ble.
This is the start of a major refactor to break up the monolithic
IOTileAdapter service and replace it with a wrapper similiar to
the CoreTools HardwareManager on top of a DeviceAdapter instance.

This will let us move beyond just pure bluetooth support to things
like MQTT and websockets as well as providing the necessary
framework for supporting things like broadcast data and realtime
device signal strength indication.